Consider using `{pins}`
For some services, it has the dual advantage of pointing to the external data source and creating version control of the data.

For my quick reading on Dropbox support, I'm not sure if it has the former

For all of this, I could well be wrong on the usefulness of pins.